requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
